+++ client/src/pages/Recipedirections.js
@@ -80,7 +80,4 @@ class listOfDirections extends Component {
     )
 }
 }
-
-
-
 }
